sound/soc/sunxi: Only disable clk when enabled to avoid warning
When a clk is not enabled it makes no sense to disable it as doing so
results in an ugly warning printed by the Kernel.

This commit changes the behavior that the sunxi_codec driver checks
if the clk is enabled before trying to disable it and thus avoiding the
warning.

Signed-off-by: Simon Eisenmann <simon@longsleep.org>